To clarify, the C client's aerospike_key_exists() was a different function prior to adding metadata. It previously had a boolean (output) parameter EXISTS which was populated with true if the record exists and false if not found. The addition of the metadata modified the contract for the function. Simply adding a metadata (output) parameter does not make it equivalent to the old operation. I was confusing the prior implementation and current implementation.
I wish I had sat down and looked at the change done to the client...so I could have given the feedback earlier...

In the old implementation, the function's contract was that it returned OK to indicate the successful execution of the function. While the EXISTS output parameter indicated the state of the record it self (FOUND vs NOT FOUND).

With the metadata change and removal of the EXISTS parameter, the contract of the function changed. No longer does it simply tell you it succeeded and a record exists, but it instead gives you data, much like the get() operation. With this change, I believe the correct change would have been to return the AEROSPIKE_ERR_RECORD_NOT_FOUND status code, just as the get() operation does. Because it is NOW effectively a get() operation without bins... which is quite different from what it was previously.