move client commands over to new JS Client class
Adds a new JS Client interface that wraps the existing C++
AerospikeClient interface. The JS Client object is responsible for
handling optional arguments. It will pass null to the corresponding
C++ client method for any optional arguments that are not present so
that the argument handling in the C++ layer can be much simplified.

The JS Client interface is also responsible to handle callbacks. It
uses a pluggable callback handler to support both the Node.js callback
semantics where the error parameter is set to null in case of a
successful operation, as well as semantics of the old Aerospike client,
where the error parameter always is a valid status object.

Likewise a new JS Query object is introduced as well that wraps the C++
Query object.

Definition of Operators and Filters are broken out into separate JS
files but with identical semantics.