fix pattern for rpm based packages
On RPM based platforms the version number of the generated *.rpm
packages does not match the version generated by pkg/version if the
version includes a subbuild number, e.g. 4.0.6-2 instead of just 4.0.6.
The *.rpm package filename uses "_" instead of "-" as the separtor, as
"-" is already used to separate the package release, e.g.

    aerospike-client-c-libev-devel-4.0.6_2-1.el7.x86_64.rpm

Relaxing the shell glob pattern to exlucde the $rev to match all
packages regardless of version format.