Add as_event_destroy_loops function.
When an external loop is used (as_event_set_external_*loop)
as_event_close_loops does not have the capacity to wait on the threads
nor to free the memory used. Thus, when one wants to use an external
event loop, the Aerospike client is only instantiable once.
The function as_destroy_loops is can be used to free the memory and
reset the global to null.
It is only safe to call is as_event_close_loops returns true.