Attempting to introduce new tend thread functionality
WIP: I'm not sure this commit is complete or self-contained.  I have some confusion over whether I should be calling as_socket_destroy() myself, or if there is a better API to use.  Similarly with respect to the metrics of the whole situation; when I close a socket, should I be updating a metric somewhere?