Tend thread to attempt to drain socket
This patch attempts to drain the socket by reading a 4KB chunk of raw
bytes.  If successful, it will replace the socket back into the
connection pool by selecting a random node and adding it there.
Otherwise, depending on the kind of error, it will either requeue the
socket for later processing in a future tend thread cycle, or it will
dispose of the socket entirely if the error is severe enough.