Factor out as_time_ctx into its own header.
Prior to this patch, it was part of as_command.h; however, as I've been
working more on this feature, I've had need to use bits and bobs of the
as_timeout_ctx structure and/or its supporting types.  This patch makes
the structure and its supporting types a separate header for easier
inclusion without having to read in all of as_command.h.