Fix divide-by-zero error, reported by Julian.
A queue needs to have at least one element; however, the minimum number
of connections defaults to zero if left unspecified.  Prior to this
patch, Julian would run tests and would see a division by zero fault.
The problem was tracked to the use of the modulo operator in C.  By
initializing the queue with the number of connections plus one, we
ensure at least one slot is available in the queue (unless you're
creating exactly 0xFFFFFFFF connections).