Introduce timeout delay logic for authenticated connections
The vast majority of the code paths leading to authenticate a connection
do not have access to an as_command structure, nor to a policy which
supports the timeout_delay setting.  As a result, I had to
"bubble up" additional context information conditionally via
as_timeout_ctx pointers (and checking to see if the pointer is
valid).  However, I believe this commit is sufficient to plumb
that information properly for those cases where timeout delay
information is available.