CLIENT-3532: Remove irrelevant code.
The code removed in this patch will probably make a come-back in later
revision when exploring the different methods of making a TLS connection are
completed and documentation thusly updated.

For now, this patch brings the code in the examples/basic_examples/connect
directory in-line with the proposed C coding documentation changes.